EJ tool

Main function that performs screening (land and water-based). Input must be an SF object! User must make this transformation.

EJfunction(
  data_type,
  facility_data,
  input_type = NULL,
  gis_option = NULL,
  buff_dist = NULL,
  threshold = NULL,
  state = NULL,
  ds_mode = NULL,
  ds_dist = NULL,
  save_option = F,
  input_name = NULL,
  attains = NULL,
  raster_data = "data/US Census Grid_SF2010_TIFF"
)

Arguments

data_type
Required. Either "landbased" or "waterbased"

facility_data
Required.

input_type
Requried. Input must be "sf" object or list of catchments (ComIDs)

gis_option
User specified method of creating buffers around areas of interest (intersect, centroid, intersection). Default is intersection.

buff_dist
Distance(s) used to create buffers (miles). Default is 1, 3, and 5 miles.

threshold
User specified threshold to represent potential concern. Default is 80%.

state
User can restrict screening to particular states. Default is to screen for entire contiguous US.

ds_mode
Set Upstream/downstream option for water-based screening. Default is downstream.

ds_dist
Set distance to examine areas upstream/downstream for water-based screening. Default is 50 miles.

input_name
Vector of names for facilities

attains
Option to pull data from the attains database. Default is FALSE.

raster_data
Dasymetric raster data. Default is to use 1kmX1km raster data from NASA's Socioeconomic Data and Applications Center (SEDAC)

Value

Examples

#===============================================================================
#=====================FEATURE 1: LAND-BASED ANALYSIS============================
#===============================================================================
# Demonstration using meat and poultry plant facility information
# How it works:
# Provide lat lons to tool and specify data type (facility_latlons)
# options to consider
# 1) gis_option. Three options available: intersect, centroid, intersection.
#    Instersection is default.
# 2) buff_dist. Radius to use around facilities
# 3) Threshold for EJ consideration. EJScreen uses 80 as default.
# 4) states. Can restrict analysis to specific states.
# bring in data for contiguous US
a1 <- EJfunction(data_type="landbased", facility_data = facilities, gis_option="centroid",
                buff_dist = 5)

#> Reading layer `EJSCREEN_StatePct' from data source 
#>   `C:\Users\atheisin\Documents\GitHub\USEPA\EJSCREENBatch\docs\reference\EJSCREEN data\EJSCREEN_2020_StatePctile.gdb' 
#>   using driver `OpenFileGDB'
#> Simple feature collection with 220333 features and 123 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -19951910 ymin: 2017837 xmax: -7254524 ymax: 11554350
#> Projected CRS: WGS 84 / Pseudo-Mercator

#> Error in tibble::rowid_to_column(., "shape_ID"): object 'facilities' not found

#===============================================================================
#=======================FEATURE 2: WATER-BASED ANALYSIS=========================
#===============================================================================
# Demonstration using random set of catchments
# How it works:
# Provide COMIDs to tool and specify data type (water_catchments)
# options to consider
# 1) ds.us.mode. Upstream/downstream option
#    mode DD is ds w/ diversions, DM is ds mainstem, UT is us w/ tributaries, UM is us mainstem.
#    Mode DD is default.
# 2) ds.us.dist. Upstream/downstream distance. 50miles is default.
# 3) buff.dist. Buffer distance around catchments in miles. 1 mile is default.
# 4) Attains. Call attains API for data? (T/F). Default is False

c <- EJfunction(data_type="waterbased", facility_data=facilities,
                input_type = 'sf', attains = F)

#> Error in tibble::rowid_to_column(., "shape_ID"): object 'facilities' not found